requirements.txt
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
SpeechRecognition==3.10.0
pyaudio==0.2.13
